refactor(crs): adding IntRange to grapher helpers instead of separate utils file
