+++ etl/grapher_helpers.py
@@ -4,6 +4,8 @@ import yaml
 import slugify
 import warnings
 import logging
+from dataclasses import dataclass, field
+
 from pathlib import Path
 
 from typing import Optional, Dict, Literal, cast, List, Any, Set, Iterable
@@ -286,3 +288,34 @@ def join_sources(sources: List[catalog.meta.Source]) -> catalog.meta.Source:
             meta[key] = sep.join(keys)
 
     return catalog.meta.Source(**meta)
+
+
+@dataclass
+class IntRange:
+    min: int
+    _min: int = field(init=False, repr=False)
+    max: int
+    _max: int = field(init=False, repr=False)
+
+    @property  # type: ignore
+    def min(self) -> int:
+        return self._min
+
+    @min.setter
+    def min(self, x: int) -> None:
+        self._min = int(x)
+
+    @property  # type: ignore
+    def max(self) -> int:
+        return self._max
+
+    @max.setter
+    def max(self, x: int) -> None:
+        self._max = int(x)
+
+    @staticmethod
+    def from_values(xs: List[int]) -> Any:
+        return IntRange(min(xs), max(xs))
+
+    def to_values(self) -> list[int]:
+        return [self.min, self.max]
